Added purge flag for cleaning up hdfs residuals

+++ evaluation/automation/file-enc/cleanup.sh
@@ -1,5 +1,15 @@
-# Cleanup intermediate files
+#!/usr/bin/env bash
+set -e
 cd "$(realpath $(dirname "$0"))"
-rm -rf ./inputs
+
+PURGE=0
+for arg in "$@"; do [[ "$arg" == "--purge" ]] && PURGE=1; done
+
 rm -rf ./outputs
-# hdfs dfs -rm -r /file-enc
+
+if [[ $PURGE -eq 1 ]]; then
+  rm -rf ./inputs
+  hdfs dfs -rm -r -f /file-enc 2>/dev/null || true
+fi
+
+echo "[cleanup] file-enc done (purge=$PURGE)"
